Removed [gs]et_installation_path to build_[gs]et
All instances of _installation_path has been changed
to build_.
This streamlines the programming and makes it a bit more robust.

Fully implemented the LUA-module type files.
TODO add a more robust variable reference in the LUA files.

TODO install all the LUA modules on the HPC and surt machines.
This will enable the use of Lmod instead of modules.

In the LUA modules we can deal with families, which is a shortcut
to "conflict". It is more robust and more intuitive.